+++ src/database/sqlite/event/event_database.py
@@ -1033,6 +1033,7 @@ class EventDatabase(MigrationDatabase):
             'SELECT `id`, `last_name`, `ratings`, `first_name`, '
             '`date_of_birth`, `year_of_birth`, `gender`, `mail`, `phone`, '
             '`comment`, `owed`, `paid`, `title`, `women_title`, `fide_id`, '
+            '`national_id`, `national_source`, '
             '`federation`, `club`, `fixed`, `check_in`, `team_id`, '
             '`team_index`, `plugin_data` FROM `player`'
         )
@@ -1054,6 +1055,8 @@ class EventDatabase(MigrationDatabase):
             title,
             women_title,
             fide_id,
+            national_id,
+            national_source,
             federation,
             club,
             fixed,
@@ -1083,6 +1086,8 @@ class EventDatabase(MigrationDatabase):
                     title=title,
                     women_title=women_title,
                     fide_id=fide_id,
+                    national_id=national_id,
+                    national_source=national_source,
                     federation=federation,
                     club=club,
                     fixed=fixed,
@@ -1110,6 +1115,8 @@ class EventDatabase(MigrationDatabase):
                 'title',
                 'women_title',
                 'fide_id',
+                'national_id',
+                'national_source',
                 'federation',
                 'club',
                 'fixed',